ve.dm.SourceConverter: Add a trailing newline to each full paragraph
Currently, the ve.dm.SourceConverter#getSourceTextFromDataRange method, which is used
to convert a linear data fragment to plain text, only inserts a newline after a paragraph
if it is followed by another paragraph node. This causes an error in the 2017 wikitext editor's
wikitable insertion tool, as the last trailing newline in the inserted wikitext is discarded,
causing the CodeMirror syntax highlight overlay to be out of sync with the editor state.

This patch reinstates the pre-Id1996d9293d1c8059040d2d9c6d93bd6b4f55ee0 behavior of
the ve.dm.SourceConverter#getSourceTextFromDataRange method by making it insert a newline
after each full paragraph it encounters in the data to be converted.

Bug: T243606
Change-Id: Ibe092f9e9423ac95890885954a1f7ce417037bd4