README: Add link to live demo
Make the README a bit more appealing to new developers from
outside the MediaWiki realm by starting off with a link to a live
demo and the API docs for its classes. For later would also be
good if there was a markdown page in the doc output that guides
devs on how to integrate VE (e.g. how the demo was made).

Also update the Gerrit link to match the URL we advertise on GitHub
mirrors for how to contribute.

Change-Id: I92c639e863084d551e4ddd204b57e194d783c332