Create a separate descriptionItemsStack for each reflist
We generate reflist diffs before the main diff so they can be
spliced into the correct part of the document. We don't want
their description items to be at the top of the stack though,
so store them in separate stacks and only add them to the main
stack when the reference list is actually rendered.

Bug: T272315
Change-Id: I373f395a2322f79545d4580005b4d587c576c6a1