ve.createDocumentFromHtml: Do not lose <style> tags in doc fragments
ve.createDocumentFromHtml always treats the input like a complete HTML
document, even though in various places we pass it document fragments.
When such fragment starts with a <meta> or <style> element, the element
is placed in the automatically generated <head> rather than <body>,
which breaks our assumptions elsewhere in the code.

We already had test cases that should have been failing because of
this bug, but the test code added the <body> itself…

Bug: T273234
Change-Id: I0759feebb7fee0fdf17e2c4589461a5ca9f536a2