Use boolean RegExp.test() instead of .match() where possible
.test() is the dedicated syntax for a boolean "does match? yes/no?"
check. .match() returns an array of matches, or null. This is just not
needed in these situations.

Also: Add missing ^ anchor to avoid possibly running unbound pattern
searches. This is not strictly necessary when we know the input is a
single character. Still I find it more robust and more obvious what
it's meant to do. Not to "search for some space somewhere" but to
check if the input _is_ whitespace.

Otherwise patterns are not touched.

Change-Id: I2d9e028c4fc11dd60926b225a13a89a2c1fd0c42