Revert "DiffElement: Collapse tree diff with no described changes"
This reverts commit 7a5a14235cfc6e0bdbf78219aedcddd7e6d480a9.

Reason for revert: Deciding to hide a node after we have
run #processQueue is too late, as we have already decided
the context to show around that node. If we want to hide a
node we have to decide this during the model diffing stage,
or add a second processing step.

Change-Id: Ie3f10c6d1648e4f8e6cca3b55cdeff33beb5ba7d