ModeledFactory: Allow candidates to suppress other candidates when matching
Currently the candidate has to be a sub-class of another candidate
to suppress it, but this can lead to strange hacks to force a
certain inheritance tree. Instead allow candidates to just specify
which other candidates they suppress.

This will be used in Citoid to allow template-specific contexts
(e.g. book/web/news) to suppress the generic context.

Bug: T310290
Change-Id: I72daa1c5d2490c71591658f186c089ab899b5683