ve.js: Update ve.now() to match optimisation of MediaWiki core
In a performance timeline for Wikipedia pageviews, I noticed that
the `ext.visualEditor.ve` module, which contains only this ve.js file
and is eagerly loaded on all page views, stands out due to accessing
`performance.timing` directly during the main JS task that implements
all the batch-loaded modules on a given article. That's a relatively
expensive DOM getter, that shows up prominently as a ~12ms task in
a 6x slowed down profile that has a PTL of 8 seconds overall.

This effectively backports my 2018 change to MW (Id621b08fbcce886).

Change-Id: I8f1808fab1185e9e9842ac93d457a0a65a0fcd23