Replace ListStorage with ConflictableStorage
ConflictableStorage handles conflicts when using localStorage
by ensuring that a set of specified keys are always overwritten
as a group when switching between instances (browser tabs).

The efficient appendToList functionality of ListStorage is
removed as just use obect storage is fast enough, even for
long transaction lists, and adds complexity that is hard
to support with ConflictableStorage, and even harder to support
with upcoming storage expiry.

Bug: T218663
Change-Id: I3c266ee93e51d1e8c4ff449b41bfe88d95b22921