Revert "Verify key before removing node from InternalList"
This reverts commit ba0714b07ed03d35c35181f80358ee8c7f53f6c9.

Reason for revert: Had an oversight when writing this patch
and discovered a bug just now. In the case of Extension:Cite,
copying a node attached to a key in the InternalList, pasting
it in a way that the key needs to be updated, and then undoing
the change ends up preventing the key from being updated.
Unsure if this is a fault with the way Extension:Cite is
written, or if this affects how InternalList is designed in
the first place. Nevertheless, this will affect the master
version of Extension:Cite so a revert is needed. Will attempt
look for a better alternative to this patch; hopefully one
that doesn't include modifying VisualEditor code. My
apologies for this.

Change-Id: I19e839d8f2f3f23f9de288479c0c0332cd90102d