Slightly increase DimensionsWidget input fields width
The widget is currently used in only 2 places:
1. In the "media settings" dialog for the "image size".
2. In the Kartographer dialog for the map size.

In both cases having slightly more space to interact with the
fields (e.g. click or tap to position the text cursor) is helpful,
esp. when the numbers are bigger.

This patch also optimizes the CSS a little to make it more
flexible/compact.

Change-Id: If94e19b2407044793f178365f338c7a1a423a465