List diffs: Fix overwriting top-level list when nested item added
appendListItem() tried to splice `listNodeData` at position 0 when
inserting a new list item in an existing list, but the data may refer
to a nested list (at unknown position), not the top-level list at
position 0.

In order to correctly handle attribute changes for the top-level list,
remove the special handling for it, and just handle it as an insertion
of a nested list (by starting `depth` at -1, which will cause the
first call to appendListItem() use `depthChange` of 1 instead of 0).

Attribute changes for nested lists may not be shown in some cases now,
but at least they don't clobber the top-level list.

Bug: T324354
Change-Id: I7c5f2db10ec2f4d38e07a05a8a1be8d90b5433d5