List diffs: Try to handle list type changes when the list node is gone
When a nested list gets un-indented, its list node disappears, and the
items get merged into the parent list. If the two lists had different
types, we had no place to show the change of the type of the
un-indented list items.

Try to show it on the list items instead. It actually mostly works?
It's only a bit problematic that the list style (bullet/number/indent)
and the definition list item style (term/definition) are both just
called 'style' in the code.

Add a test for list type change to definition list (indent) to restore
ve.dm.DefinitionListNode.js test coverage (previously covered only by
one of the changed tests).

Change-Id: I62adf509e65b68a56f343975b53e4a39a1b124cd