ve.ce.Surface: Check dataTransfer.files is non-empty
Whenever we are checking dataTransfer.files what we really
mean is to check if it is non-empty, so check the length as
well.

Since we dropped IE support we probably don't need to check
if the files property is set, but that would require some
more rigourous testing to check it wouldn't cause an error.

Change-Id: I2d22fb485a3361615f71d15724345925e71e88ef