Make sure InternalList sorting is stable both ways
I run into this while working with the code. While this might be
unreachable under real-world circumstances right now it feels odd to
assume that only the first function argument can be undefined, but
the second cannot. There is nothing here that would give such a
guarantee, as far as I can see.

This might be responsible for certain subtle user-facing bugs. But
I currently don't have a specific Phabricator ticket in mind to link
this to, sorry.

Change-Id: I62fa0d84ace67246a3b345c4a7b9d2688ee4d641