Fix incomplete comparison in TextState.getChunks
We found this while working on something unrelated, see I3b624f9.

The problem here is that the comparison always fails. The add()
function is actually called with type being undefined. But the
other side of the comparison is already normalized to the string
"text".

The effect is that the "merging" described in the function
documentation never happens.

There is a chance this might be responsible for subtle bugs.
Unfortunatelly I wasn't able to find a scenario where this would
make a difference. Still I find it important to fix it.

Change-Id: I45cf87f0a69577ba250735d7d3122e7c7223bf27