Add named, documented ve.dm.InternalListNodeGroup class
As of now this doesn't really do anything. The individual objects
are the same as before, with the same properties. The only effect
of this patch is that it gives us a place for documentation and to
possibly add methods (in later patches).

Note that some of these properties are already accessed from other
codebases, so changing them needs to be done slow and carefully
anyway. Being able to document them as e.g. @deprecated should make
this much easier.

Bug: T397145
Bug: T397395
Change-Id: I1778c9174c31c871ac028c56700ab0f3bd47ed1a