Move getUniqueListKey method into ve.dm.InternalListNodeGroup
I find this method especially interesting because it is the only
place where two (!) of the properties in the InternalListNodeGroup
class are ever used. This is clearly private stuff that should be
hidden to not confuse users of this data structure.

In this patch I'm intentionally moving the code without making any
other change.

Bug: T397395
Change-Id: Ifd0a2d38b22cc232bb8d7cdc8454593093083e16