Remove InternalList's dependency on document.buildingNodeTree
I think there is no good reason to do something different in one
mode vs. the other. The only difference is that one of the methods
is a little bit cheaper because it avoids re-sorting the nodes. But
that's not terribly expensive anyway. The sorting is done based on
`node.getRange().start`. And it's only done for the list of
references in a document, which is typically something like 10, or
at most a few 100.

To make the sorting even cheaper we can use getOffset() directly
instead of getRange(). This is guaranteed to result in the same
order – even if start and end of a range are flipped. It doesn't
matter which position inside of a range we use when all we care
about is which range comes first. The ranges we work with here are
impossible to overlap.

Also make some of the new InternalListNodeGroup methods @private.

Bug: T397501
Bug: T424632
Change-Id: I2a1460016a4aaa4e8966fc005f3b46ccca7c041b