Optimize ve.ce.Document hasClass() check for performance
This code was introduced in patch I8b833fe in 2016. I found it a
little confusing because of the repeated `&& s` at the end. After
playing around a bit I realized it does the same `.nodeType` checks
over and over again. While these are not expensive it can still add
up when the loop is long.

By moving the `.nodeType` checks up the majority of the loop cvan
be skipped in case it doesn't match anyway.

Change-Id: Ie928077c8e5842464d976d11299958ad9154f773