Rewrite ve.escapeHtml utility function with modern syntax
The function was written in a way where it was compiling itself as a
function. This extra indirection was added in 2015 via Idbf4128.

I think this was probably not necessary even back in 2015.

The approach with a map instead of a `switch` is short and fast and
recommended in several places.

Change-Id: I90e9c7b7a89b609bd847df0bc2f19f009f78f95d