Use hasAttribute to read the most mininal data-ve-ignore=""
Direct follow-up for Ibd80ea5, as suggested.

Note this attribute is set in several codebases, namely the VE and
Cite extensions. It's always set to "1" at the moment. But the
exact value never mattered. We can as well represent this with an
empty string and achieve the same.

With the new version of the ModelFromDomConverter released we should
be free to change this value freely in all places where it's set.
Both old and new serializations will work identical.

Change-Id: I1547f1d04b3e176aa78ab500e134427d8dd2da6f