Remove uses of transparent pixel in CSS
In both cases it is being used to keep an empty paragraph open.
We can use an NBSP instead, which is only a few px wide so
unlikely to change the layout.

In CBN (e.g. paragraph nodes) it doesn't affect cursor placement
or rendering. Also we use inline slugs now so I'm not sure this
is triggered anymore.

In visual diffs having an NBSP is fine too.

Change-Id: I889948de2ec6f7d513dde25a67c53d203d97bfd5