Add all necessary public methods to ve.dm.InternalListNodeGroup
This is a re-submission of I4ebdebf and I9df413a and I6648c44
(here squashed). The first two got temporarily reverted to unblock
I9f2a08d. Turns out there was indeed a small mistake in the new
removeNode method. This is only called from Cite. The mistake was
hard to spot because of this. Now fixed.

I do my best to keep the new code in InternalListNodeGroup as
isolated as possible. It should not know anything about the document.
It might mention "references" in comments but the code should not
need to know anything about how references actually work.

Some details worth mentioning:
* With this patch it's almost possible to hide all properties in the
  class as private. I already mark them with a (soft) @private to
  make it easier to migrate existing code.
* The only exception is when "appendNodeWithKnownIndex" accepts an
  index from a previous serialization. This needs to stay for now.
  All other situations that exposed the index before are not needed
  any more.
* The old code uses a property "registeredListKey" that is set in
  Cite. VE and the VE lib shouldn't need to know anything about it.
  This patch solves this circular dependency.

Bug: T397145
Change-Id: I46d9e9a5f6f319e3cf7482e9a78353b6408625b6