Two more fixes for the (existing) InternalList behavior
Direct follow-up for a few more minor mistakes in I46d9e9a.
Unfortunately the relevant code is only called from Cite and only
tested from Cite.

The wrong comparison was a human-error I made while re-working
parts of the code. It's not well covered with tests and probably
not even relevant for real-world scenarios.

The other problem is that not only appendNodeWithKnownIndex()
can be called with any index out of order, but
insertNodeInDocumentOrder() can as well. One of the two methods
might be obsolete now. But this is something I want to have a look
at later, after this is fixed.

Bug: T397145
Change-Id: I0eebc62e50713c7d8dbd649458b1a1080bcb5175