Fix incomplete word-wrapping in ve.ui.LinearContextItem
1. This replaces the deprecated alias "word-wrap" with the new
standard "overflow-wrap".

2. Instead of "break-word" use "anywhere". The two are almost the
same, "but soft wrap opportunities introduced by the word break are
NOT considered when calculating min-content intrinsic sizes." This
is a problem in this context. The edit button still disappears at
the right when the label on the left is to long. The edit button is
unreachable then. This should never happen.

https://developer.mozilla.org/en-US/docs/Web/CSS/overflow-wrap

Bug: T404690
Change-Id: Icdce626ed22a486bb7c8aa042b2779fff064a5eb