Fix loop in InternalList iterating one time to much
This patch is about loops that run backwards. One of them started
at the wrong position and was always doing one iteration to much.

I have seen the short `for ( i = length ; i--; )` syntax in many,
many places in core and extensions. I prefer it because it is short
and there is not much that can go wrong.

Change-Id: I3cedd1bace9cee065e42ef53438216fe6c5e6fb1