Inline iLen (and such) variables in for() loops
I realize this is some kind of established code style in this
codebase. Still it keeps confusing me. I feel like it only has
disadvantages:

* It makes the code longer and harder to read. It's an extra level
of indirection. In some cases the variable really is needed inside
of the loop, but more often it's not. It's hard to tell the
difference and when it matters. The variable names are often
different. Sometimes it's even a lowercase "l" which is hard to
read and to distinguish from e.g. "i".

* Accessing a .length property is not expensive. Modern JS runtime
engines are very optimized. They understand the default structure of
a for() loop very well and do everything they can to run it as fast
as possible. It's more likely that the extra variable does the
opposite and makes the loop slower.

What does the editing team think?

Change-Id: I813dcb2abffeb115c41a45febcbbc0713ce5fe25