ve.ui.FragmentInspector.test: Speed up slow test
All the fragment inspector tests were run in series so that
a single surface could be reused. Each test took ~150ms but
createing a new surface only took ~6ms, so it is much faster
to try to run the tests in parallel, creating a separate
surface for each test.

As we have some global state that is mocked, run desktop then
mobile tests separately.

Total run time is ~500ms locally, down from ~3,200ms.

Change-Id: I931587afa32e6d13c3b973b96033f06231acbe7f