Maintain keyIndexes when rebuilding InternalList
While coding a feature which looks at ve.dm.InternalItemNode in the
absence of a ve.dm.MWReferenceNode, we discovered that the
InternalList will lose some of its internal indexes after cloning.
Rebuilding the index during addNode will preserve the necessary
information.

Note how a comment in the clone method explains: "Most properties
don't need to be copied, because addNode() will be invoked when the
new document tree is built." That's why this change is in the addNode
method but still relevant for clone.

Bug: T406869
Change-Id: I9e501ce1228edfefc01a1980b7670ad591f7ac69