Qunit: Improve references tests fixture
I make sure the fixture is complete and more realistic.  The
attributes left out should never be missing and it also seems that
currently there's no test that expecting these missing attributes.

This is a preparation to better show the effect of the change that
rebuilds the keyIndexes property in InternalList.

In addition to that I'm adding two assertions showing the output
of the current test helper method that creates the example
documents.

See I9e501ce1228edfefc01a1980b7670ad591f7ac69

Bug: T408535
Change-Id: I9d9a60f68684eef7e54498fea4a8a7e709ddf970