Qunit: Use `references` fixture for InternalList.clone() test
I think the fixure using references might be better to test some
of the aspects of cloning an InternalList.  It seems to make no
difference for the current test setup.

I'm also documenting the behavior of the test setup regarding the
keyIndexes property.  Mostly as preparation to show the effect of
the change that rebuilds the keyIndexes property in InternalList.

See I9e501ce1228edfefc01a1980b7670ad591f7ac69

Bug: T408535
Change-Id: I68d72a75fbad641c5f8679354b8eff842939cc24