Maintain keyIndexes when rebuilding InternalList
While coding a feature which looks at ve.dm.InternalItemNode in the
absence of a ve.dm.MWReferenceNode, we discovered that the
InternalList will lose some of its internal indexes after cloning.
Rebuilding the index during addNode will preserve the necessary
information.

Note how a comment in the clone method explains: "Most properties
don't need to be copied, because addNode() will be invoked when the
new document tree is built." That's why this change is in the addNode
method but still relevant for clone.

This is a resubmit of I9e501ce1228edfefc01a1980b7670ad591f7ac69
I did a lot of digging into the issue that let to the revert of
the original patch, see T408535.  Due to these investigations I'm
convinced that this patch is not harmful, but in contrary might
even set a few things right.

The changes in the tests do show the implications on the test
setups and the effects when cloning a document and rebuilding
it's tree.

Bug: T408535
Bug: T406869
Change-Id: I58fb936ee068c262da0c02b9325dbb41febbb1af