Add safe-guard to unstable VisualDiff.freezeInternalListIndices
This code was rewritten not long ago in I46d9e9a. Usually it should
be impossible for the loop to contain a key that then cannot be found
by the following `getAllReuses` call. But while demoing a draft patch
in today's demo time we run into this.

This is a safe guard until we can identify the actual problem. We
will continue working on the diff code in one of our next sprints.

Bug: T397145
Change-Id: I1bb4b2be868c5895fa485b823666ef9ca2b3ba69