Block user from adding whitespace-only comments
We found this while working on T296575. Note that comment nodes are
edited with a WhitespacePreservingTextInputWidget that preserves the
whitespace before and after the text in a <!-- comment --> node. The
VE user cannot see this whitespace and cannot edit it.

The current behavior is confusing: It's possible to create a comment
node that contains nothing but whitespace. But when you try to edit
it the text field appears empty and the node gets removed because the
content is empty now.

New behavior: Don't add empty comments in the first place.

We believe this is more consistent and less surprising.

Change-Id: Iad35e9a17bce69f254d6de23d75c8ea390203391